Updating the overview UI initial state
