+++ components/apimgt/org.wso2.carbon.apimgt.internal.service/swagger.json
@@ -966,6 +966,11 @@
           "example" : "APIProduct",
           "description" : "Type of the API."
         },
+        "status" : {
+          "type" : "string",
+          "example" : "APIProduct",
+          "description" : "Type of the API."
+        },
         "isDefaultVersion" : {
           "type" : "boolean",
           "example" : false,
