+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.admin.v1/.openapi-generator/admin-api.yaml.sha256
@@ -1,1 +1,1 @@
-4f101301789b29a0e8dd4dca211ea6ddad1aff3a7bc954dc6e86511460a21299
+0bf2fcc7b7e357121a5446d57a03ff40ea027a55e010250f48f6842d10a96716
