+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/devportal/site/public/locales/en.json
@@ -130,9 +130,10 @@
   "Apis.Details.InfoBar.keyManagers": "Key Managers",
   "Apis.Details.InfoBar.list.context": "Context",
   "Apis.Details.InfoBar.list.endpoint": "Endpoint",
+  "Apis.Details.InfoBar.list.provider": "By : ",
   "Apis.Details.InfoBar.list.tags": "Tags",
   "Apis.Details.InfoBar.list.tags.not": "Not Tagged",
-  "Apis.Details.InfoBar.list.version": "Version",
+  "Apis.Details.InfoBar.list.version": "Version : ",
   "Apis.Details.InfoBar.security.api_key": "Api Key",
   "Apis.Details.InfoBar.security.basic": "Basic",
   "Apis.Details.InfoBar.security.oauth2": "OAuth2",
@@ -173,6 +174,7 @@
   "Apis.Details.StarRatingSummary.not.rated": "Not Rated",
   "Apis.Details.StarRatingSummary.user": "user",
   "Apis.Details.StarRatingSummary.users": "users",
+  "Apis.Details.Topics.error.404": "Resource not found",
   "Apis.Details.TryOutConsole.access.token.tooltip": "You can use your existing Access Token or you can generate a new Test Key.",
   "Apis.Details.index.all.apis": "ALL APIs",
   "Apis.Details.index.comments": "Comments",
